update HIRES pipeline
